requirements.txt
Django==2.1
djangorestframework==3.9
requests==2.21